Fold test support into integration suite
