+++ crates/shelterwood/Cargo.toml
@@ -15,7 +15,7 @@ tokio-util.workspace = true
 tracing.workspace = true
 
 [dev-dependencies]
-shelterwood-test-support = { path = "../test-support" }
+tokio = { workspace = true, features = ["test-util"] }
 
 [[test]]
 name = "integration"
