+++ crates/shelterwood/tests/acceptance/sidecar.rs
@@ -7,14 +7,12 @@ use std::{
     time::Duration,
 };
 
+use crate::common::{ReleaseGate, policy::never, poll_until};
 use shelterwood::{
     Actor, ActorOnceDef, ChildState, Context, ExitError, ExitKind, ExitResult, Readiness,
     ReadinessDeadline, Shutdown, StartOrShutdownError, StopContext, SubtreeOnceDef, TaskDef,
     TaskRef, Tree,
 };
-use shelterwood_test_support::{ReleaseGate, poll_until};
-
-use crate::common::policy::never;
 
 type JournalEvent = (usize, &'static str, &'static str);
 type JournalEntries = Arc<Mutex<Vec<JournalEvent>>>;
