--- crates/test-support/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "shelterwood-test-support"
-version = "0.1.0"
-edition = "2024"
-rust-version = "1.97.1"
-license = "MIT OR Apache-2.0"
-publish = false
-
-[dependencies]
-tokio = { workspace = true, features = ["macros", "test-util"] }
